include comments from create pr Signed-off-by: Ruirui Zhang <mariazrr@amazon.com>
