+++ plugins/workload-management/src/main/java/org/opensearch/plugin/wlm/action/TransportDeleteQueryGroupAction.java
@@ -10,10 +10,9 @@ package org.opensearch.plugin.wlm.action;
 
 import org.opensearch.action.support.ActionFilters;
 import org.opensearch.action.support.HandledTransportAction;
-import org.opensearch.cluster.metadata.QueryGroup;
 import org.opensearch.common.inject.Inject;
 import org.opensearch.core.action.ActionListener;
-import org.opensearch.plugin.wlm.action.service.Persistable;
+import org.opensearch.plugin.wlm.service.QueryGroupPersistenceService;
 import org.opensearch.tasks.Task;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.TransportService;
@@ -21,12 +20,12 @@ import org.opensearch.transport.TransportService;
 /**
  * Transport action for delete QueryGroup
  *
- * @opensearch.internal
+ * @opensearch.experimental
  */
 public class TransportDeleteQueryGroupAction extends HandledTransportAction<DeleteQueryGroupRequest, DeleteQueryGroupResponse> {
 
     private final ThreadPool threadPool;
-    private final Persistable<QueryGroup> queryGroupPersistenceService;
+    private final QueryGroupPersistenceService queryGroupPersistenceService;
 
     /**
      * Constructor for TransportDeleteQueryGroupAction
@@ -35,7 +34,7 @@ public class TransportDeleteQueryGroupAction extends HandledTransportAction<Dele
      * @param transportService - a {@link TransportService} object
      * @param actionFilters - a {@link ActionFilters} object
      * @param threadPool - a {@link ThreadPool} object
-     * @param queryGroupPersistenceService - a {@link Persistable} object
+     * @param queryGroupPersistenceService - a {@link QueryGroupPersistenceService} object
      */
     @Inject
     public TransportDeleteQueryGroupAction(
@@ -43,7 +42,7 @@ public class TransportDeleteQueryGroupAction extends HandledTransportAction<Dele
         TransportService transportService,
         ActionFilters actionFilters,
         ThreadPool threadPool,
-        Persistable<QueryGroup> queryGroupPersistenceService
+        QueryGroupPersistenceService queryGroupPersistenceService
     ) {
         super(DeleteQueryGroupAction.NAME, transportService, actionFilters, DeleteQueryGroupRequest::new);
         this.threadPool = threadPool;
@@ -53,6 +52,7 @@ public class TransportDeleteQueryGroupAction extends HandledTransportAction<Dele
     @Override
     protected void doExecute(Task task, DeleteQueryGroupRequest request, ActionListener<DeleteQueryGroupResponse> listener) {
         String name = request.getName();
-        threadPool.executor(ThreadPool.Names.GENERIC).execute(() -> queryGroupPersistenceService.delete(name, listener));
+        threadPool.executor(ThreadPool.Names.GENERIC)
+            .execute(() -> queryGroupPersistenceService.deleteInClusterStateMetadata(name, listener));
     }
 }
