--- plugins/workload-management/src/main/java/org/opensearch/plugin/wlm/action/rest/package-info.java
@@ -1,12 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * Package for the rest classes for QueryGroup CRUD operations
- */
-package org.opensearch.plugin.wlm.action.rest;
